requirements.txt
flask==3.0.3
flask-socketio==5.3.6
python-dotenv==1.0.1
eventlet==0.35.2
gunicorn==21.2.0
